Replace Heightmap with FilteredMap
Rename FilteredMap to Heightmap. Perform associated object renames. Replace FilteredMapUtilities with MapFilterFactory to better reflect file's status as a factory for float functions.
