--- Editor.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: 44a367875532db1448e4218ae2b387ac
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
