--- Editor/HeightmapDrawer.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: f0ff17c747a286a4f8661554caa99b0d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
