--- Scenes/DemoPropertyDrawer.meta
@@ -1,8 +1,0 @@
-fileFormatVersion: 2
-guid: f9e7079039b926644a34b14e704b0f53
-folderAsset: yes
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
