--- Scenes/DemoPropertyDrawer/DemoPropertyDrawer.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: 453442294d726cf4b8d8eaa6afdcab1f
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
