--- Scenes/DemoPropertyDrawer/DemoPropertyHeightmap.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9691363b7814c2a44b4a56bc6b67321d
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
