--- Scenes/HeightmapDemoScene.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: d67a4e1ead385e340b397f666e754226
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
