--- Scenes/Scripts/HeightmapToTexture.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 989ae7256dc80c043a121545ede0a84c
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
