--- Scenes/Scripts/HeightmapWorld.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 147d686b30466f749abfe8339f1179f1
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
