--- Scenes/WorldScaleHeightmap.unity.meta
@@ -1,7 +1,0 @@
-fileFormatVersion: 2
-guid: e743964552756444ab90a018ac1721ad
-DefaultImporter:
-  externalObjects: {}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
