--- Scripts/FilteredMap.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 27f8a8061099ba143af57a0b7b505cad
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
