--- Scripts/MapDoodadColliderFlag.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 9f50d82ce03702f438e73e21ce4d6791
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
